backend/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.21.1
requests==2.31.0
SQLAlchemy==2.0.0
psycopg2-binary==2.9.5
alembic==1.10.0
redis==4.5.0
loguru==0.7.0
python-multipart==0.0.6
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
slowapi==0.1.8
feedparser==6.0.10
aiofiles==23.1.0
geopy==2.3.0
beautifulsoup4==4.12.2
lxml==4.9.3
rq==1.15.1
rq-scheduler==0.13.0

# NLP and Geocoding dependencies
spacy==3.6.0
spacy-lookups-data==1.0.3
langdetect==1.0.9
openai==1.0.0
nltk==3.8.1
shapely==2.0.1
tenacity==8.2.2

# Additional dependencies
python-multipart==0.0.6
pytz==2023.3
python-dateutil==2.8.2
httpx==0.24.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.10.1
